Implemented Search
A simple search for courses.  Added a form to index.html, a route path in routerApp and some logic in there to attach a listener to the form button and extract the search term.  Then an option object is created and passed to courseList via the load method, courseList.sync modified to extract the action, subject and search filter.